requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.1.19=h06a4308_1
certifi=2020.12.5=py38h06a4308_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
gmp=6.2.1=h2531618_2
gnutls=3.6.5=h71b1129_1002
intel-openmp=2020.2=254
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h2531618_0
matplotlib=3.4.1=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.3.0=py38h54f3939_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=hbb512f6_0
ninja=1.10.2=py38hff7bd54_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
olefile=0.46=py_0
opencv-python=4.5.1.48=pypi_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
pandas=1.2.3=pypi_0
pillow=8.2.0=py38he98fc37_0
pip=21.0.1=py38h06a4308_0
protobuf=3.15.7=pypi_0
pyparsing=2.4.7=pypi_0
python=3.8.8=hdb3f193_4
python-dateutil=2.8.1=pypi_0
pytorch=1.8.1=py3.8_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pypi_0
readline=8.1=h27cfd23_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.2=pypi_0
setuptools=52.0.0=py38h06a4308_0
six=1.15.0=py38h06a4308_0
sklearn=0.0=pypi_0
sqlite=3.35.4=hdfb4753_0
tensorboardx=2.2=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
torchvision=0.9.1=py38_cu102
tqdm=4.60.0=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
